fix-#55: PlaylistRepository에서 컬러 정보 관련 메서드 ColorInfoRepository로 이동
