+++ domain/src/main/java/pyc/domain/repository/PlaylistRepository.kt
@@ -29,13 +29,4 @@ interface PlaylistRepository {
         playlistId: Int,
         changedOrderList: List<Pair<Int, Int>>
     ): Completable
-
-    fun updateColorInfo(
-        id: Int,
-        colorInfo: ColorInfo
-    ): Completable
-
-    fun getColorInfo(
-        id: Int,
-    ): Flowable<ColorInfo>
 }
