+++ domain/src/main/java/pyc/domain/usecase/GetColorInfoUseCase.kt
@@ -2,12 +2,12 @@ package pyc.domain.usecase
 
 import io.reactivex.Flowable
 import pyc.domain.model.ColorInfo
-import pyc.domain.repository.PlaylistRepository
+import pyc.domain.repository.ColorInfoRepository
 import javax.inject.Inject
 
 class GetColorInfoUseCase @Inject constructor(
-    private val playlistRepository: PlaylistRepository
+    private val colorInfoRepository: ColorInfoRepository
 ) {
-    operator fun invoke(id: Int) : Flowable<ColorInfo> =
-        playlistRepository.getColorInfo(id)
+    operator fun invoke(id: Int): Flowable<ColorInfo> =
+        colorInfoRepository.getColorInfo(id)
 }
