+++ domain/src/main/java/pyc/domain/usecase/UpdateColorInfoUseCase.kt
@@ -1,14 +1,15 @@
 package pyc.domain.usecase
 
 import pyc.domain.model.ColorInfo
+import pyc.domain.repository.ColorInfoRepository
 import pyc.domain.repository.PlaylistRepository
 import javax.inject.Inject
 
 class UpdateColorInfoUseCase @Inject constructor(
-    private val playlistRepository: PlaylistRepository
+    private val colorInfoRepository: ColorInfoRepository
 ) {
     operator fun invoke(
         id: Int,
         colorInfo: ColorInfo
-    ) = playlistRepository.updateColorInfo(id, colorInfo)
+    ) = colorInfoRepository.updateColorInfo(id, colorInfo)
 }
